+++ HaishinKit/Sources/Codec/VTSessionOptionKey.swift
@@ -46,6 +46,8 @@ struct VTSessionOptionKey: Codable, RawRepresentable {
     static let maxH264SliceBytes = VTSessionOptionKey(rawValue: kVTCompressionPropertyKey_MaxH264SliceBytes as String)
     static let maxFrameDelayCount = VTSessionOptionKey(rawValue: kVTCompressionPropertyKey_MaxFrameDelayCount as String)
     static let encoderID = VTSessionOptionKey(rawValue: kVTVideoEncoderSpecification_EncoderID as String)
+    static let prioritizeEncodingSpeedOverQuality = VTSessionOptionKey(rawValue: kVTCompressionPropertyKey_PrioritizeEncodingSpeedOverQuality as String)
+    static let referenceBufferCount = VTSessionOptionKey(rawValue: kVTCompressionPropertyKey_ReferenceBufferCount as String)
 
     @available(iOS 16.0, tvOS 16.0, macOS 13.0, *)
     static let constantBitRate = VTSessionOptionKey(rawValue: kVTCompressionPropertyKey_ConstantBitRate as String)